requirements.txt
Flask==2.0.1
openpyxl==3.0.9
Pillow==9.0.0
openpyxl-image-loader==1.0.5
gunicorn==20.1.0
werkzeug==2.0.2